progen_code/requirements.txt
absl-py==0.9.0
astor==0.8.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.0
cloudpickle==0.8.1
cycler==0.10.0
Cython==0.29.6
decorator==4.4.1
defusedxml==0.6.0
entrypoints==0.3
gast==0.3.3
google-pasta==0.1.8
grpcio==1.27.1
h5py==2.9.0
importlib-metadata==1.5.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.0.1
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.0.3
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.16.2
openslide-python==1.1.1
pandas==0.24.2
pandocfilters==1.4.2
parso==0.6.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==6.0.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
ptyprocess==0.6.0
pycurl==7.43.0
Pygments==2.5.2
pygobject==3.20.0
pyparsing==2.3.1
pyrsistent==0.15.7
python-apt==1.1.0b1+ubuntu0.16.4.2
python-dateutil==2.8.0
pytz==2018.9
pyzmq==18.1.1
qtconsole==4.6.0
scikit-learn==0.20.0
scipy==1.2.1
Send2Trash==1.5.0
six==1.12.0
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-gpu==1.14.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
torch==1.3.0
torchvision==0.2.2.post3
tornado==6.0.3
tqdm==4.42.1
traitlets==4.3.3
wcwidth==0.1.8
webencodings==0.5.1
Werkzeug==1.0.0
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==2.2.0
